feat(module) Move the `validate` function into `Module.validate`.
This patch creates the `Module` class, with so far a single static
method `validate`.
